requirements.txt
fastapi
uvicorn
python-dotenv
GitPython
PyGithub
google-generativeai
pydantic
aiofiles
httpx